Extract interface for working with single and multiple transactions
